migracao_postgres

+++ db/migrate/20160109171313_add_open_balance_to_projects.rb
@@ -1,5 +1,5 @@
 class AddOpenBalanceToProjects < ActiveRecord::Migration
   def change
-    add_column :projects, :open_balance, :double, :default => 0
+    add_column :projects, :open_balance, :decimal, :default => 0
   end
 end
