+++ db/migrate/20160114001857_add_interest_rate_to_project.rb
@@ -1,5 +1,5 @@
 class AddInterestRateToProject < ActiveRecord::Migration
   def change
-    add_column :projects, :interest_rate, :double
+    add_column :projects, :interest_rate, :decimal
   end
 end
