+++ db/migrate/20160114114112_add_installment_to_project.rb
@@ -1,5 +1,5 @@
 class AddInstallmentToProject < ActiveRecord::Migration
   def change
-    add_column :projects, :installment, :double
+    add_column :projects, :installment, :decimal
   end
 end
